# A bit of FillTessellator optimization