--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,6 +13,10 @@ license = "MIT/Apache-2.0"
 name = "lyon"
 path = "src/lib.rs"
 
+# Uncomment this when profiling.
+#[profile.release]
+#debug = true
+
 [dependencies]
 
 lyon_tessellation = { version = "0.7.3", path = "tessellation/" }