I spent some time profiling and optimizing the tessellator this weekend.

before:
```
test logo_events_and_tess           ... bench:  15,668,913 ns/iter (+/- 173,931)
test logo_tess_no_curve             ... bench:   4,509,290 ns/iter (+/- 63,883)
test logo_tess_no_intersection      ... bench:   7,758,496 ns/iter (+/- 639,328)
test logo_tess_only                 ... bench:   9,252,560 ns/iter (+/- 984,362)
```


after:
```
test logo_events_and_tess           ... bench:  12,028,170 ns/iter (+/- 153,927)
test logo_tess_no_curve             ... bench:   3,337,986 ns/iter (+/- 21,012)
test logo_tess_no_intersection      ... bench:   5,540,060 ns/iter (+/- 65,196)
test logo_tess_only                 ... bench:   6,798,144 ns/iter (+/- 79,672)
```